fix subtle bug around coinductive predicates
In particular, we can't accept a chain of reasoning
that is based on hypotheses. In terms of co-lp, the
way to think of this is that proving things from the
hypotheses is inductive, whereas proving things with an
impl is coinductive. This seems to imply some kind of argument
about a cycle of impls never being infinite, given that
there are a fine number of impls and a finite set of types,
but I have to think more on that!