stop removing variables from the environment
Now that we have introduced subtyping, we will sometimes have
variables whose values are not known precisely in the environment,
so we can't be sure to have substituted them all out.

We probably *could* find values for them and do the substitution,
but this version, where the state of inference persists,
is closer to how a real implementation would work.