extend higher-ranked types in the same way
The test `wf--hrty.rkt` shows the implications.
The user experience is kind of inconsistent as a result,
but it's compatible with what rustc does today from
I can tell.

There are limits to how much we can change this while
keeping a simple, local translation (though we could
make some of these cases error if desired).