introduce coherence-mode, but don't use it
Now in coherence we can prove things in 'coherence-mode',
but right now that's the same as proving any other way.
This change revelead that we were invoking `is-predicate?`
on both goals and hypotheses. It happened to work before
because the grammars were sufficiently overlapping, but
it doesn't now, so we have to get a bit more particular.