automatically reject variable names
When one variant is tagged as `#[variable]`,
we now automatically reject variable names from
other variants. This means we can use `#[precedence]`
just to encode true *precedence* (i.e., of operators).
This was important since there was a conflict where
variables wanted to be the same *precedence* as other
identifiers in a logical sense.