modify `#[variable]` to take kind explicitly
The `variable` parser method is now simpler
and just returns a variable; this also lets
`Variable` implement `Parse`.

We now have the helper `variable_of_kind`.

This is all useful in Dada which wants to parse variables
in slightly more flexible ways and in particular
*doesn't* want to force them to be downcasted.

Overall it seems more obvious.